requirements.txt
certifi==2020.6.20
chardet==3.0.4
contextvars==2.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
lxml==4.6.1
python-docx==0.8.10
rfc3986==1.4.0
sniffio==1.2.0
